webgpu/shader/execution: Shuffle harness code
Promote the builtin test runner to `src/webgpu/shader/execution/expression.ts`,  and tweak the interface so it can be used for any kind of expression.

Move comparision utilities out to `src/webgpu/util/compare.ts`.
Move constants out to `src/webgpu/util/constants.ts`.
