+++ src/webgpu/api/operation/vertex_state/correctness.spec.ts
@@ -874,7 +874,7 @@ g.test('vertex_buffer_used_multiple_times_interleaved')
     const alignedFormatByteSize = align(formatByteSize, 4);
 
     // Create data for a single vertex buffer with many attributes, that will be split between
-    // many vertexbuffers set at different offsets.
+    // many vertex buffers set at different offsets.
 
     // In this test we want to test using the same vertex buffer for multiple different attributes.
     // For example if vbCount is 3, we will create a vertex buffer containing the following data:
