+++ src/webgpu/shader/execution/builtin/all.spec.ts
@@ -5,8 +5,7 @@ Execution Tests for the 'all' builtin function
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
 import { False, True, TypeBool, TypeVec, vec2, vec3, vec4 } from '../../../util/conversion.js';
-
-import { run } from './builtin.js';
+import { builtin, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -79,5 +78,5 @@ e: vecN<bool> all(e): bool Returns true if each component of e is true. (OpAll)
     };
     const overload = overloads[t.params.overload];
 
-    run(t, 'all', [overload.type], TypeBool, t.params, overload.cases);
+    run(t, builtin('all'), [overload.type], TypeBool, t.params, overload.cases);
   });
