+++ src/webgpu/shader/execution/builtin/atan.spec.ts
@@ -4,10 +4,11 @@ Execution Tests for the 'atan' builtin function
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
+import { ulpThreshold } from '../../../util/compare.js';
+import { kBit, kValue } from '../../../util/constants.js';
 import { f32, f32Bits, TypeF32 } from '../../../util/conversion.js';
 import { biasedRange, linearRange } from '../../../util/math.js';
-
-import { Case, Config, kBit, kValue, run, ulpThreshold } from './builtin.js';
+import { builtin, Case, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -71,5 +72,5 @@ TODO(#792): Decide what the ground-truth is for these tests. [1]
 
     const cfg: Config = t.params;
     cfg.cmpFloats = ulpThreshold(4096);
-    run(t, 'atan', [TypeF32], TypeF32, cfg, cases);
+    run(t, builtin('atan'), [TypeF32], TypeF32, cfg, cases);
   });
