+++ src/webgpu/shader/execution/builtin/cos.spec.ts
@@ -4,10 +4,10 @@ Execution Tests for the 'cos' builtin function
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
+import { absThreshold } from '../../../util/compare.js';
 import { f32, TypeF32 } from '../../../util/conversion.js';
 import { linearRange } from '../../../util/math.js';
-
-import { absThreshold, Case, Config, run } from './builtin.js';
+import { builtin, Case, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -38,5 +38,5 @@ TODO(#792): Decide what the ground-truth is for these tests. [1]
 
     const cfg: Config = t.params;
     cfg.cmpFloats = absThreshold(2 ** -11);
-    run(t, 'cos', [TypeF32], TypeF32, cfg, cases);
+    run(t, builtin('cos'), [TypeF32], TypeF32, cfg, cases);
   });
