+++ src/webgpu/shader/execution/builtin/countLeadingZeros.spec.ts
@@ -4,9 +4,8 @@ Execution Tests for the 'countLeadingZeros' builtin function
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
-import { i32, i32Bits, TypeI32, u32, TypeU32, u32Bits } from '../../../util/conversion.js';
-
-import { Config, run } from './builtin.js';
+import { TypeU32, u32Bits, u32, TypeI32, i32Bits, i32 } from '../../../util/conversion.js';
+import { builtin, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -33,7 +32,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
   )
   .fn(async t => {
     const cfg: Config = t.params;
-    run(t, 'countLeadingZeros', [TypeU32], TypeU32, cfg, [
+    run(t, builtin('countLeadingZeros'), [TypeU32], TypeU32, cfg, [
       // Zero
       { input: u32Bits(0b00000000000000000000000000000000), expected: u32(32) },
 
@@ -163,7 +162,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
   )
   .fn(async t => {
     const cfg: Config = t.params;
-    run(t, 'countLeadingZeros', [TypeI32], TypeI32, cfg, [
+    run(t, builtin('countLeadingZeros'), [TypeI32], TypeI32, cfg, [
       // Zero
       { input: i32Bits(0b00000000000000000000000000000000), expected: i32(32) },
