+++ src/webgpu/shader/execution/builtin/firstTrailingBit.spec.ts
@@ -5,8 +5,7 @@ Execution Tests for the 'firstTrailingBit' builtin function
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
 import { i32, i32Bits, TypeI32, u32, TypeU32, u32Bits } from '../../../util/conversion.js';
-
-import { Config, run } from './builtin.js';
+import { builtin, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -34,7 +33,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
   )
   .fn(async t => {
     const cfg: Config = t.params;
-    run(t, 'firstTrailingBit', [TypeU32], TypeU32, cfg, [
+    run(t, builtin('firstTrailingBit'), [TypeU32], TypeU32, cfg, [
       // Zero
       { input: u32Bits(0b00000000000000000000000000000000), expected: u32(-1) },
 
@@ -165,7 +164,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
   )
   .fn(async t => {
     const cfg: Config = t.params;
-    run(t, 'firstTrailingBit', [TypeI32], TypeI32, cfg, [
+    run(t, builtin('firstTrailingBit'), [TypeI32], TypeI32, cfg, [
       // Zero
       { input: i32Bits(0b00000000000000000000000000000000), expected: i32(-1) },
