+++ src/webgpu/shader/execution/builtin/insertBits.spec.ts
@@ -15,8 +15,7 @@ import {
   vec4,
   TypeVec,
 } from '../../../util/conversion.js';
-
-import { Config, run } from './builtin.js';
+import { builtin, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -97,7 +96,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
       0b01010101010101010101010101010101
     );
 
-    run(t, 'insertBits', [T, T, TypeU32, TypeU32], T, cfg, [
+    run(t, builtin('insertBits'), [T, T, TypeU32, TypeU32], T, cfg, [
       { input: [all_0, all_0, u32(0), u32(32)], expected: all_0 },
       { input: [all_0, all_0, u32(1), u32(10)], expected: all_0 },
       { input: [all_0, all_0, u32(2), u32(5)], expected: all_0 },
