+++ src/webgpu/shader/execution/builtin/inversesqrt.spec.ts
@@ -4,10 +4,11 @@ Execution Tests for the 'inverseSqrt' builtin function
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
+import { ulpThreshold } from '../../../util/compare.js';
+import { kBit, kValue } from '../../../util/constants.js';
 import { f32, f32Bits, TypeF32 } from '../../../util/conversion.js';
 import { biasedRange, linearRange } from '../../../util/math.js';
-
-import { Case, Config, kBit, kValue, run, ulpThreshold } from './builtin.js';
+import { builtin, Case, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -47,5 +48,5 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
 
     const cfg: Config = t.params;
     cfg.cmpFloats = ulpThreshold(2);
-    run(t, 'inverseSqrt', [TypeF32], TypeF32, cfg, cases);
+    run(t, builtin('inverseSqrt'), [TypeF32], TypeF32, cfg, cases);
   });
