+++ src/webgpu/shader/execution/builtin/log2.spec.ts
@@ -4,10 +4,11 @@ Execution Tests for the 'log2' builtin function
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { GPUTest } from '../../../gpu_test.js';
+import { absThreshold, ulpThreshold } from '../../../util/compare.js';
+import { kBit, kValue } from '../../../util/constants.js';
 import { f32, f32Bits, TypeF32 } from '../../../util/conversion.js';
 import { biasedRange, linearRange } from '../../../util/math.js';
-
-import { absThreshold, Case, Config, kBit, kValue, run, ulpThreshold } from './builtin.js';
+import { builtin, Case, Config, run } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -48,5 +49,5 @@ TODO(#792): Decide what the ground-truth is for these tests. [1]
       }
       return ulpThreshold(3)(got, expected);
     };
-    run(t, 'log2', [TypeF32], TypeF32, cfg, cases);
+    run(t, builtin('log2'), [TypeF32], TypeF32, cfg, cases);
   });
