+++ src/webgpu/shader/execution/builtin/select.spec.ts
@@ -20,8 +20,7 @@ import {
   vec3,
   vec4,
 } from '../../../util/conversion.js';
-
-import { run, CaseList } from './builtin.js';
+import { run, CaseList, builtin } from '../expression.js';
 
 export const g = makeTestGroup(GPUTest);
 
@@ -121,7 +120,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
 
     run(
       t,
-      'select',
+      builtin('select'),
       [overload.type, overload.type, TypeBool],
       overload.type,
       t.params,
@@ -228,7 +227,7 @@ https://github.com/gpuweb/cts/blob/main/docs/plan_autogen.md
 
     run(
       t,
-      'select',
+      builtin('select'),
       [tests.dataType, tests.dataType, tests.boolType],
       tests.dataType,
       t.params,
